bot/requirements.txt
fastapi
uvicorn
langchain-groq
python-dotenv
langchain-mcp-adapters
langgraph
langchain
langchain-community
langchain-core